+++ CHANGELOG.md
@@ -4,6 +4,7 @@
         - the global flag `--infile-list` accepts stdin "-" now.
     - `seqkit split`: forgot to add extension for --two-pass. [#332](https://github.com/shenwei356/seqkit/issues/332)
     - `seqkit faidx`: allow empty lines at the end of sequences.
+    - `seqkit stats`: fix compute Q1 and Q3 of sequence length for one record. [#353](https://github.com/shenwei356/seqkit/issues/353)
 - [SeqKit v2.3.1](https://github.com/shenwei356/seqkit/releases/tag/v2.3.1) - 2022-09-22
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v2.3.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v2.3.1)
     - `seqkit grep/locate`: fix bug of FMIndex building for empty sequences. [#321](https://github.com/shenwei356/seqkit/issues/321)